requirements.txt
Flask
CouchDB
requests
pycryptodome
websockets
